+++ src/main/java/insane96mcp/iguanatweaksreborn/command/ITRCommand.java
@@ -4,6 +4,7 @@ import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.arguments.FloatArgumentType;
 import insane96mcp.iguanatweaksreborn.entity.ITRFallingBlockEntity;
 import insane96mcp.iguanatweaksreborn.module.sleeprespawn.tiredness.TirednessHandler;
+import insane96mcp.iguanatweaksreborn.module.world.weather.Weather;
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.commands.Commands;
 import net.minecraft.commands.arguments.EntityArgument;
@@ -12,6 +13,7 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.EnchantmentHelper;
+import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraftforge.registries.ForgeRegistries;
 
@@ -43,6 +45,12 @@ public class ITRCommand {
                             player.getInventory().add(enchantedBook);
                             return 1;
                         }))
+                .then(Commands.literal("foggy_weather")
+                        .then(Commands.literal("clear")
+                                .executes(context -> {
+                                    Weather.clearFoggyWeather(context.getSource().getServer().getLevel(Level.OVERWORLD));
+                                    return 1;
+                                })))
                 .then(Commands.literal("test")
                         .executes(context -> {
                             for (int i = 0; i < 3; i++) {