+++ src/main/java/insane96mcp/iguanatweaksreborn/network/message/FoggySync.java
@@ -2,6 +2,7 @@ package insane96mcp.iguanatweaksreborn.network.message;
 
 import insane96mcp.iguanatweaksreborn.module.world.weather.ClientWeather;
 import insane96mcp.iguanatweaksreborn.module.world.weather.Foggy;
+import insane96mcp.iguanatweaksreborn.module.world.weather.WeatherSavedData;
 import insane96mcp.iguanatweaksreborn.network.NetworkHandler;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.server.level.ServerPlayer;
@@ -11,38 +12,32 @@ import net.minecraftforge.network.NetworkEvent;
 import java.util.function.Supplier;
 
 public class FoggySync {
-	private final int foggyTimer;
-	private final int foggyTargetTime;
-	private final Foggy currentFoggy;
-	private final Foggy targetFoggy;
+	private WeatherSavedData.FoggyData foggyData;
 
-	public FoggySync(int foggyTimer, int foggyTargetTime, Foggy currentFoggy, Foggy targetFoggy) {
-		this.foggyTimer = foggyTimer;
-		this.foggyTargetTime = foggyTargetTime;
-		this.currentFoggy = currentFoggy;
-		this.targetFoggy = targetFoggy;
+	public FoggySync(WeatherSavedData.FoggyData foggyData) {
+		this.foggyData = foggyData;
 	}
 
 	public static void encode(FoggySync pkt, FriendlyByteBuf buf) {
-		buf.writeInt(pkt.foggyTimer);
-		buf.writeInt(pkt.foggyTargetTime);
-		buf.writeEnum(pkt.currentFoggy);
-		buf.writeEnum(pkt.targetFoggy);
+		buf.writeInt(pkt.foggyData.timer);
+		buf.writeInt(pkt.foggyData.targetTime);
+		buf.writeEnum(pkt.foggyData.current);
+		buf.writeEnum(pkt.foggyData.target);
 	}
 
 	public static FoggySync decode(FriendlyByteBuf buf) {
-		return new FoggySync(buf.readInt(), buf.readInt(), buf.readEnum(Foggy.class), buf.readEnum(Foggy.class));
+		return new FoggySync(new WeatherSavedData.FoggyData(buf.readInt(), buf.readInt(), buf.readEnum(Foggy.class), buf.readEnum(Foggy.class)));
 	}
 
 	public static void handle(final FoggySync message, Supplier<NetworkEvent.Context> ctx) {
 		ctx.get().enqueueWork(() -> {
-			ClientWeather.updateFoggy(message.foggyTimer, message.foggyTargetTime, message.currentFoggy, message.targetFoggy);
+			ClientWeather.updateFoggy(message.foggyData);
 		});
 		ctx.get().setPacketHandled(true);
 	}
 
-	public static void sync(ServerPlayer player, int foggyTimer, int foggyTargetTime, Foggy currentFoggy, Foggy targetFoggy) {
-		Object msg = new FoggySync(foggyTimer, foggyTargetTime, currentFoggy, targetFoggy);
+	public static void sync(ServerPlayer player, WeatherSavedData.FoggyData foggyData) {
+		Object msg = new FoggySync(foggyData);
 		NetworkHandler.CHANNEL.sendTo(msg, player.connection.connection, NetworkDirection.PLAY_TO_CLIENT);
 	}
 }